requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.3
-e git+ssh://git@github.com/PhilippDahlinger/calabi_yau_metrics.git@5b5ac9bc0a02a590189f4eac363175581ea88974#egg=calabi_yau_metrics
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
dm-tree==0.1.8
docker-pycreds==0.4.0
exceptiongroup==1.2.1
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
hydra-core==1.3.2
idna==3.7
ipykernel==6.29.5
ipython==8.23.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.12.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
ml-dtypes==0.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.14.0
sentry-sdk==2.8.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.14.0
termcolor==2.4.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.5.0
urllib3==2.2.2
wandb==0.17.4
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.14.1
